+++ src/Telegram.BotAPI/BotAPI/Available Methods/restrictChatMember.cs
@@ -37,7 +37,7 @@ namespace Telegram.BotAPI.AvailableMethods
         /// <param name="untilDate">Optional. Date when restrictions will be lifted for the user, unix time. If user is restricted for more than 366 days or less than 30 seconds from the current time, they are considered to be restricted forever.</param>
         /// <exception cref="BotRequestException">Thrown when a request to Telegram Bot API got an error response.</exception>
         /// <exception cref="ArgumentNullException">Thrown when a required parameter is null.</exception>
-        public static bool RestrictChatMember(this BotClient bot, long chatId, int userId, AvailableTypes.ChatPermissions permissions, [Optional] uint untilDate)
+        public static bool RestrictChatMember(this BotClient bot, long chatId, long userId, AvailableTypes.ChatPermissions permissions, [Optional] uint untilDate)
         {
             if (bot == default)
             {
@@ -86,7 +86,7 @@ namespace Telegram.BotAPI.AvailableMethods
         /// <param name="cancellationToken">The cancellation token to cancel operation.</param>
         /// <exception cref="BotRequestException">Thrown when a request to Telegram Bot API got an error response.</exception>
         /// <exception cref="ArgumentNullException">Thrown when a required parameter is null.</exception>
-        public static async Task<bool> RestrictChatMemberAsync(this BotClient bot, object chatId, int userId, AvailableTypes.ChatPermissions permissions, [Optional] uint untilDate, [Optional] CancellationToken cancellationToken)
+        public static async Task<bool> RestrictChatMemberAsync(this BotClient bot, object chatId, long userId, AvailableTypes.ChatPermissions permissions, [Optional] uint untilDate, [Optional] CancellationToken cancellationToken)
         {
             if (bot == default)
             {
